Telas de banco de questões e visualizar questão | Método de salvar, editar e remover questões.

+++ src/main/java/com/example/repository/DisciplinaRepository.java
@@ -1,11 +1,12 @@
 package com.example.repository;
 
-import com.example.model.Disciplina;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository; 
 
+import com.example.model.Disciplina; 
+
 @Repository 
-public interface DisciplinaRepository extends JpaRepository<Disciplina, Integer> {
+public interface DisciplinaRepository extends JpaRepository<Disciplina, Long> {
     
   
 }
